Restructure the src filesystem a bit
