+++ rmf_traffic/src/rmf_traffic/geometry/Box.cpp
