+++ rmf_traffic/src/rmf_traffic/geometry/Circle.cpp
