+++ rmf_traffic/src/rmf_traffic/geometry/ConvexShape.cpp
