+++ rmf_traffic/src/rmf_traffic/geometry/Shape.cpp
