+++ rmf_traffic/src/rmf_traffic/geometry/ShapeInternal.hpp
