+++ rmf_traffic/src/rmf_traffic/geometry/SimplePolygon.cpp
