+++ rmf_traffic/src/rmf_traffic/track/VehicleTraits.cpp
